extracted Configuration models to their own file added static default configuration values
