+++ Shared/Models/MobSession.swift
@@ -10,11 +10,11 @@ import Foundation
 struct MobSession {
     var teamMembers: [TeamMember]
     var isActive = false
-    var breakLengthInSeconds = Configuration(value: 5 * 60, maxValue: 60 * 30, isTimeValue: true, label: "Break Length", color: "MobOrange")
-    var numberOfRotationsBetweenBreaks = Configuration(value: 5 * 60, maxValue: 10 * 60, isTimeValue: false, label: "Rounds Between Breaks", color: "MobYellowGreen")
+    var breakLengthInSeconds = Configuration.defaulBreakLengthInSeconds
+    var numberOfRotationsBetweenBreaks = Configuration.defaultNumberOfRotationsBetweenBreaks
     
     init() {
-        if let teamMemberData: [TeamMember] = JSONUtility.read(from: Constants.teamMemberNamesPath) {
+        if let teamMemberData: [TeamMember] = try? JSONUtility.read(from: Constants.teamMemberNamesPath) {
             teamMembers = teamMemberData
         } else {
             teamMembers = []
@@ -26,37 +26,6 @@ struct TeamMember: Identifiable, Equatable, Codable {
     var id = UUID()
     var name: String
     var role: Role
-}
-
-struct Configurations: Codable {
-    var rotationLength: Configuration
-    var breakLengthInSeconds: Configuration
-    var numberOfRotationsBetweenBreaks: Configuration
-}
-
-struct Configuration: Identifiable, Codable {
-    var id = UUID()
-    var value: Int
-    let maxValue: Int
-    let isTimeValue: Bool
-    let label: String
-    let color: String
-    var progress: Double = 0
-    var angle: Double
-    
-    var formattedValue: Int {
-        (Int(progress * Double(maxValue)) / 60) + 1
-    }
-    
-    init(value: Int, maxValue: Int, isTimeValue: Bool, label: String, color: String) {
-        self.value = value
-        self.maxValue = maxValue
-        self.isTimeValue = isTimeValue
-        self.label = label
-        self.color = color
-        progress = Double(value) / Double(maxValue)
-        angle = Double(progress * Double(360))
-    }
 }
 
 // MARK: Test Data
